separate menu component
